requirements.txt
fastapi
gspread
google-auth
uvicorn
